create logic
