+++ client/src/redux/slices/userSlice/thunks.js
@@ -49,8 +49,9 @@ export const postUser = (obj) => {
       await axios.post("/api/v1/users", {
         ...obj,
       });
+      console.log("Usuario creado");
     } catch (error) {
-      console.log(error.message);
+      console.error(error.response);
     }
   };
 };
