+++ client/src/views/AdminSection/AdminSection.jsx
@@ -1,7 +1,7 @@
 import { ProSidebarProvider } from "react-pro-sidebar";
 import Sidebar from "./AdminGlobals/Sidebar";
 import { Route, Routes} from "react-router-dom";
-import AdminUsers from "../../components/AdminSection/AdminUsers";
+import AdminUsers from "../../components/AdminSection/AdminUsers/AdminUsers";
 
 export default function AdminSection() {
